move code to reuse it in logical opt
