+++ ydb/core/kqp/opt/physical/kqp_opt_phy_impl.h
@@ -31,23 +31,6 @@ NYql::NNodes::TDqStage ReplaceStageArg(NYql::NNodes::TDqStage stage, size_t inpu
 NYql::NNodes::TDqStage ReplaceTableSourceSettings(NYql::NNodes::TDqStage stage, size_t inputIndex,
     NYql::NNodes::TKqpReadRangesSourceSettings settings, NYql::TExprContext& ctx);
 
-enum ESortDirection : ui32 {
-    None = 0,
-    Forward = 1,
-    Reverse = 2,
-    Unknown = 4,
-};
-
-using ESortDirectionRaw = std::underlying_type<ESortDirection>::type;
-
-inline ESortDirection operator|(ESortDirection a, ESortDirection b) {
-    return ESortDirection(static_cast<ESortDirectionRaw>(a) | static_cast<ESortDirectionRaw>(b));
-}
-
-inline ESortDirection operator|=(ESortDirection& a, ESortDirection b) { return (a = a | b); }
-
-ESortDirection GetSortDirection(const NYql::NNodes::TExprBase& sortDirections);
-
 NYql::TExprNode::TPtr MakeMessage(TStringBuf message, NYql::TPositionHandle pos, NYql::TExprContext& ctx);
 
 } // NKikimr::NKqp::NOpt
